Add move_arm overlay that fixes a segfault.
There are no checks on the return value of the trajectory filter call, so when
it fails move_arm tries to get references to first() and last() elements of an
empty vector, hence the segfault.

The warehouse version of move_arm has a fix for this, but not the plain move_arm
used by us.